updating to SeedRoles module
